+++ pele/potentials/_hs_wca_cpp.pyx
@@ -16,9 +16,9 @@ cdef extern from "pele/hs_wca.h" namespace "pele":
     cdef cppclass  cHS_WCANeighborList "pele::HS_WCANeighborList":
         cHS_WCANeighborList(_pele.Array[long] & ilist, double eps, double sca, _pele.Array[double] radii) except +    
     cdef cppclass  cHS_WCAFrozen "pele::HS_WCAFrozen":
-        cHS_WCAFrozen(double eps, double sca, _pele.Array[double] radii, _pele.Array[double]& reference_coords, _pele.Array[long]& frozen_dof) except +
+        cHS_WCAFrozen(double eps, double sca, _pele.Array[double] radii, _pele.Array[double]& reference_coords, _pele.Array[size_t]& frozen_dof) except +
     cdef cppclass  cHS_WCAPeriodicFrozen "pele::HS_WCAPeriodicFrozen":
-        cHS_WCAPeriodicFrozen(double eps, double sca, _pele.Array[double] radii, double* boxvec, _pele.Array[double]& reference_coords, _pele.Array[long]& frozen_dof) except +
+        cHS_WCAPeriodicFrozen(double eps, double sca, _pele.Array[double] radii, double* boxvec, _pele.Array[double]& reference_coords, _pele.Array[size_t]& frozen_dof) except +
 
 cdef class HS_WCA(_pele.BasePotential):
     """define the python interface to the c++ HS_WCA implementation
@@ -55,13 +55,13 @@ cdef class HS_WCAFrozen(_pele.BasePotential):
         if boxvec is None:
             self.periodic = False
             self.thisptr = <_pele.cBasePotential*>new cHS_WCAFrozen(eps, sca, _pele.Array[double](<double*> radii.data, radii.size), _pele.Array[double](<double *> reference_coords.data, reference_coords.size),
-                        _pele.Array[long](<long *> frozen_dof.data, frozen_dof.size) )
+                        _pele.Array[size_t](<size_t *> frozen_dof.data, frozen_dof.size) )
         else:
             self.periodic = True
             bv = np.array(boxvec, dtype=float)
             assert bv.size == 3
             self.thisptr = <_pele.cBasePotential*>new cHS_WCAPeriodicFrozen(eps, sca, _pele.Array[double](<double*> radii.data, radii.size), <double*> bv.data, _pele.Array[double](<double *> reference_coords.data, reference_coords.size),
-                        _pele.Array[long](<long *> frozen_dof.data, frozen_dof.size) )
+                        _pele.Array[size_t](<size_t *> frozen_dof.data, frozen_dof.size) )
 
 cdef class HS_WCANeighborList(_pele.BasePotential):
     """define the python interface to the c++ HS_WCA implementation
