+++ source/pele/lj.h
@@ -106,7 +106,7 @@ namespace pele {
 
     class LJFrozen : public FrozenPotentialWrapper<LJ> {
         public:
-            LJFrozen(double C6, double C12, Array<double> & reference_coords, Array<long int> & frozen_dof)
+            LJFrozen(double C6, double C12, Array<double> & reference_coords, Array<size_t> & frozen_dof)
                 : FrozenPotentialWrapper< LJ > 
                   ( new LJ(C6, C12), reference_coords, frozen_dof ) {}
     };
